--- pkg/analyzer/testdata/src/_crypto/crypto.go
@@ -1,5 +1,0 @@
-package _crypto
-
-func MD4() {
-	_ = "MD4" // want `can use crypto.MD4.String\(\) instead "MD4"`
-}
