+++ pkg/analyzer/testdata/src/a/crypto/crypto.go
@@ -1,0 +1,5 @@
+package crypto
+
+func MD4() {
+	_ = "MD4" // want `"MD4" can be replaced by crypto.MD4.String\(\)`
+}
